pyproject.toml
[project]
name = "realtime-bus-location"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "kafka-python>=2.1.5",
    "uvicorn>=0.34.0",
]
